adjust navbar header
