+++ src/components/LineCurve/LineCurve.tsx
@@ -1,19 +1,11 @@
-// import LineChatType from "../../interfaces/LineChatType";
-import linecurve from "../../data/linecurve";
-import "./LineCurve.css";
-
-const lineCurveChatArr = linecurve.data.chatDetails;
-const LineChatCurves = lineCurveChatArr.map((item: any, index) => {
-  return (
-    <div className="totalCountChat">
-      <div className="totalCount">{item.totalCount}</div>
-      <div className="totalCountTitle">{item.totalCountTitle}</div>
-    </div>
-  );
-});
+import LineCurver from "./LineCurver";
 
 const LineCurve = () => {
-  return <div className="chat">{LineChatCurves}</div>;
+  return (
+    <>
+      <LineCurver />
+    </>
+  );
 };
 
 export default LineCurve;
