+++ src/components/Navbar/Navbar.module.css
@@ -7,7 +7,7 @@
 
 .topNavBar {
   display: flex;
-  justify-content: space-around;
+  justify-content: space-between;
   height: 4rem;
   align-items: center;
 }
@@ -18,6 +18,7 @@
   margin-right: 30px;
   margin-left: 30px;
 }
+
 .topNavBar ul .navItem {
   list-style-type: none;
   padding: 0.65rem 1.1rem;
@@ -32,6 +33,7 @@
 .navItem {
   display: flex;
   align-items: center;
+  font-weight: bold;
 }
 
 .navItemPosition {
@@ -82,3 +84,4 @@
   display: flex;
   align-items: center;
 }
+
